feat(router): implement routing
issue #5
