+++ mobile/App.js
@@ -6,40 +6,28 @@ import {
   View
 } from 'react-native';
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: '#F5FCFF',
-  },
-  welcome: {
-    fontSize: 20,
-    textAlign: 'center',
-    margin: 10,
-  },
-  instructions: {
-    textAlign: 'center',
-    color: '#333333',
-    marginBottom: 5,
-  },
-});
+import ScarletScreen from './SignIn';
+import GrayScreen from './SignUp';
+import {Router, Scene} from 'react-native-router-flux';
 
-const App = () => (
-  <View style={styles.container}>
-    <Text style={styles.welcome}>
-      Welcome to React Native!
-    </Text>
-    <Text style={styles.instructions}>
-      To get started, edit /mobile/App.js
-    </Text>
-    <Text style={styles.instructions}>
-      Press Cmd+R to reload,{'\n'}
-      Cmd+D or shake for dev menu
-    </Text>
-  </View>
-);
-
+const App = () => {
+    return (
+        <Router>
+          <Scene key="root">
+            <Scene key="scarlet"
+                   component={ScarletScreen}
+                   title="SignIn"
+                   initial
+            />
+            <Scene
+                key="gray"
+                component={GrayScreen}
+                title="SignUp"
+            />
+          </Scene>
+        </Router>
+    );
+}
 App.displayName = 'App';
 
 export default App;
