+++ package.json
@@ -100,6 +100,8 @@
     "react": "^15.4.1",
     "react-dom": "^15.4.1",
     "react-helmet": "^3.2.2",
+    "react-native": "0.41.2",
+    "react-native-router-flux": "^3.38.0",
     "react-redux": "^4.4.5",
     "react-router": "^2.7.0",
     "react-router-redux": "^4.0.0",
@@ -115,13 +117,10 @@
     "url-loader": "^0.5.7",
     "webpack": "^2.2.1",
     "webpack-dev-middleware": "^1.9.0",
-    "webpack-hot-middleware": "^2.14.0",
-    "react-native": "0.41.2"
+    "webpack-hot-middleware": "^2.14.0"
   },
   "jest": {
     "preset": "react-native"
   },
-  "babel": {
-
-  }
+  "babel": {}
 }
